Extract and rename
